Simplifies the workflow for fitting and training a knn model

Simplifies the workflow for fitting and training a knn model

model_workflow(data, num_folds, range_neighbors, response)

Arguments

data

The dataset to be split into training and testing sets

num_folds

The desired number of folds to use for cross validation

range_neighbors

A sequence of the desired values of k

response

The response variable

Value

a fitted model

Examples

data <- data.frame(
a = rnorm(1000),
b = rnorm(1000),
cultivar = sample(c(1,2,3), 1000, replace = TRUE)
)
num_folds <- 5
range_neighbors <- seq(1, 20)
response <- "cultivar"
model_workflow(data, num_folds, range_neighbors, response)
#> $cmat
#> Confusion Matrix and Statistics
#> 
#>           Reference
#> Prediction  1  2  3
#>          1 34 31 26
#>          2 32 25 25
#>          3 18 32 28
#> 
#> Overall Statistics
#>                                          
#>                Accuracy : 0.3466         
#>                  95% CI : (0.2879, 0.409)
#>     No Information Rate : 0.3506         
#>     P-Value [Acc > NIR] : 0.5761         
#>                                          
#>                   Kappa : 0.0194         
#>                                          
#>  Mcnemar's Test P-Value : 0.5068         
#> 
#> Statistics by Class:
#> 
#>                      Class: 1 Class: 2 Class: 3
#> Sensitivity            0.4048   0.2841   0.3544
#> Specificity            0.6587   0.6503   0.7093
#> Pos Pred Value         0.3736   0.3049   0.3590
#> Neg Pred Value         0.6875   0.6272   0.7052
#> Prevalence             0.3347   0.3506   0.3147
#> Detection Rate         0.1355   0.0996   0.1116
#> Detection Prevalence   0.3625   0.3267   0.3108
#> Balanced Accuracy      0.5317   0.4672   0.5319
#> 
#> $metrics
#>           Accuracy Precision    Recall F1_Score
#> Accuracy 0.3466135 0.3458263 0.3477611 0.346459
#>