Read data at a given directory and factor columns if needed

Read data at a given directory and factor columns if needed

read_and_factor(input_dir, factor_vars = NULL)

Arguments

input_dir

The directory where the file to be read is located

factor_vars

a single string or vector of strings representing variables to factor

Value

a dataframe

Examples

read_and_factor('~/predictcultivar/tests/testthat/data/unfactored_populated_data.csv', "cultivar")
#> New names:
#> • `` -> `...1`
#> Rows: 200 Columns: 4
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (1): cultivar
#> dbl (3): ...1, total_phenols, alcohol
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 200 × 4
#>     ...1 total_phenols alcohol cultivar
#>    <dbl>         <dbl>   <dbl> <fct>   
#>  1     1        0.380   1.70   a       
#>  2     2       -0.0876 -1.66   c       
#>  3     3       -1.56   -1.82   b       
#>  4     4        0.878  -0.594  b       
#>  5     5        0.504   0.774  c       
#>  6     6       -1.39   -0.266  a       
#>  7     7        0.131   0.622  a       
#>  8     8       -1.06   -1.98   a       
#>  9     9       -0.102   0.0532 a       
#> 10    10        0.0522 -0.123  b       
#> # ℹ 190 more rows
read_and_factor('~/predictcultivar/tests/testthat/data/unfactored_populated_data.csv',
c("cultivar","alcohol"))
#> New names:
#> • `` -> `...1`
#> Rows: 200 Columns: 4
#> ── Column specification ────────────────────────────────────────────────────────
#> Delimiter: ","
#> chr (1): cultivar
#> dbl (3): ...1, total_phenols, alcohol
#> 
#> ℹ Use `spec()` to retrieve the full column specification for this data.
#> ℹ Specify the column types or set `show_col_types = FALSE` to quiet this message.
#> # A tibble: 200 × 4
#>     ...1 total_phenols alcohol            cultivar
#>    <dbl>         <dbl> <fct>              <fct>   
#>  1     1        0.380  1.70348905747801   a       
#>  2     2       -0.0876 -1.66396043347481  c       
#>  3     3       -1.56   -1.82306934441291  b       
#>  4     4        0.878  -0.594040817835244 b       
#>  5     5        0.504  0.773876136880341  c       
#>  6     6       -1.39   -0.266003095170955 a       
#>  7     7        0.131  0.621808536061493  a       
#>  8     8       -1.06   -1.97569256443609  a       
#>  9     9       -0.102  0.0531861584714699 a       
#> 10    10        0.0522 -0.122718098591163 b       
#> # ℹ 190 more rows